Refactoring, change BaseModel to have necessary DynamoDb annotations.
