+++ src/main/java/gov/cdc/izgateway/xform/model/AccessControl.java
@@ -9,7 +9,7 @@ import java.util.UUID;
 
 @Getter
 @Setter
-public class AccessControl implements BaseModel {
+public class AccessControl extends BaseModel {
     private UUID id;
     @NotNull(message = "Access Control - User is required")
     @ValidUser(message = "User ID must reference an existing and active user")
